+++ api/kilosim/LightPattern.h
