+++ api/kilosim/random.hpp
